restructure the code of ltr
